move other enpoints to separate pkg
