+++ backend/endpoints/endpoints.go
@@ -19,6 +19,13 @@ func RegisterEndpoints(router *gin.Engine) {
 	router.GET("/shows", getShows)
 	router.GET("/shows/:id", getShow)
 	router.GET("/shows/count", getShowCount)
+	router.GET("/ratings/best", getBestRating)
+	router.GET("/episodes/:parentTconst", getShowEpisodes)
+	router.POST("/users", createUser)
+	router.GET("/users/:id", getUser)
+	router.GET("/users/all", getAllUsers)
+	router.DELETE("/users/:id", deleteUser)
+	router.GET("/validateUser", validateUser)
 }
 
 func CloseDB() {
